requirements.txt
alembic==1.15.2
azure-core==1.32.0
azure-identity==1.15.0
black==24.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coverage==7.4.3
cryptography==43.0.3
et_xmlfile==2.0.0
flake8==7.0.0
Flask==3.0.2
flask-cors==6.0.1
Flask-DebugToolbar==0.16.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
mccabe==0.7.0
msal==1.25.0
msal-extensions==1.1.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.4
numpy==2.3.0
openpyxl==3.1.5
packaging==24.2
pandas==2.3.0
pathspec==0.12.1
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.0.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse2==1.7.1
pytz==2025.2
reportlab==4.1.0
requests==2.31.0
simplejson==3.20.1
six==1.17.0
SQLAlchemy==2.0.40
sqlite3-to-mysql==2.4.0
tabulate==0.9.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
Unidecode==1.3.8
urllib3==2.3.0
Werkzeug==3.0.1
WTForms==3.2.1
